requirements.txt
fastapi
uvicorn[standard]
playwright
chromadb
python-dotenv
google-generativeai
openai-whisper
python-multipart